requirements.txt
# Optional: matplotlib if you want higher-quality plots
# matplotlib
python-pptx
